Extract more components into separate files
