+++ src/app/components/EditorialSummary/EditorialSummary.tsx
@@ -196,17 +196,15 @@ export const EditorialSummary = withStyles(styles)(
       return (
         <>
           <div className={classes.articleText}>
-            {nodes
-              .filter(isRootBlock)
-              .map(node => (
-                <Block
-                  key={node.id}
-                  node={node}
-                  nodes={nodes}
-                  referencesMap={this.references}
-                  onSourceClick={this.onSourceClick}
-                />
-              ))}
+            {nodes.filter(isRootBlock).map(node => (
+              <Block
+                key={node.id}
+                node={node}
+                nodes={nodes}
+                referencesMap={this.references}
+                onSourceClick={this.onSourceClick}
+              />
+            ))}
             <footer>
               <Typography color="textSecondary" component="small">
                 This article was written by {author}
